Interest group UI fixes
