+++ src/features/interest-groups/components/index.ts
@@ -6,3 +6,5 @@
 
 export * from "./interest-group-card";
 export * from "./interest-groups-grid";
+export * from "./interest-groups-client";
+export * from "./interest-group-detail-client";
